Changed data type to in32 for ID and utilized Upsert method for Store method instead of Insert

+++ pkg/entity/driver.go
@@ -1,7 +1,7 @@
 package entity
 
 type Driver struct {
-	Id       uint8   `json:"id" bson:"id"`
+	Id       int32   `json:"id" bson:"id"`
 	Lat      float64 `json:"latitude" bson:"latitude"`
 	Long     float64 `json:"longitude" bson:"longitude"`
 	Accuracy float64 `json:"accuracy" bson:"accuracy"`
